+++ pkg/validator/crd_decoder.go
@@ -1,4 +1,4 @@
-package validatorfactory
+package validator
 
 import (
 	"errors"
