+++ pkg/validator/factory_test.go
@@ -1,4 +1,4 @@
-package validatorfactory
+package validator
 
 import (
 	"os"
