+++ pkg/validator/schema_patcher.go
@@ -1,4 +1,4 @@
-package validatorfactory
+package validator
 
 import (
 	"path/filepath"
