+++ pkg/validator/structural.go
@@ -1,4 +1,4 @@
-package validatorfactory
+package validator
 
 import (
 	"fmt"
