+++ pkg/validator/testdata/crds/cronjob.yaml
